requirements.txt
colorama==0.4.6
Delorean==1.0.0
disnake==2.9.2
pymongo==4.12.0
python-dotenv==1.1.0
Requests==2.32.3
